@@ -20,16 +20,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-builders/custom-webpack": "^7.0.2",
+    "@angular-builders/custom-webpack": "^8.4.1",
     "@angular-builders/dev-server": "^7.3.1",
-    "@angular/animations": "^7.2.15",
-    "@angular/common": "^7.2.2",
-    "@angular/core": "^7.2.2",
-    "@angular/forms": "^7.2.2",
-    "@angular/http": "^7.2.2",
-    "@angular/platform-browser": "^7.2.2",
-    "@angular/platform-browser-dynamic": "^7.2.2",
-    "@angular/router": "^7.2.2",
+    "@angular/animations": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/forms": "^8.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
     "@ionic-native/app-version": "^5.15.1",
     "@ionic-native/core": "^5.15.1",
     "@ionic-native/device": "^5.15.1",
@@ -48,16 +47,16 @@
     "@ngx-translate/core": "^11.0.1",
     "@ngx-translate/http-loader": "^4.0.0",
     "@project-sunbird/client-services": "3.6.12",
-    "@project-sunbird/common-consumption": "3.6.9",
+    "@project-sunbird/common-consumption-v8": "3.6.4",
     "@project-sunbird/content-player": "3.6.1",
     "@project-sunbird/discussions-ui": "1.0.11",
     "@project-sunbird/sb-styles": "0.0.7",
-    "@project-sunbird/sunbird-pdf-player": "0.1.2",
+    "@project-sunbird/sunbird-pdf-player-v8": "3.6.0",
     "@project-sunbird/sunbird-sdk": "3.6.5",
     "code-push": "^3.0.1",
     "com.jjdltc.cordova.plugin.zip": "https://github.com/swayangjit/jjdltc-cordova-plugin-zip.git",
     "com.telerik.plugins.nativepagetransitions": "^0.6.5",
-    "common-form-elements": "1.0.8",
+    "common-form-elements-v8": "2.0.4",
     "cordova-android": "^9.0.0",
     "cordova-plugin-advanced-http": "^2.5.1",
     "cordova-plugin-android-downloadmanager": "git+https://github.com/Sunbird-Ed/sb-cordova-plugin-downloadmanager.git",
@@ -105,7 +104,7 @@
     "lodash-es": "^4.17.20",
     "ng-recaptcha": "^5.0.0",
     "ngx-chips": "2.0.2",
-    "rxjs": "^6.5.3",
+    "rxjs": "^6.6.3",
     "sb-cordova-plugin-customtabs": "git+https://github.com/Sunbird-Ed/sb-cordova-plugin-customtabs.git",
     "sb-cordova-plugin-db": "git+https://github.com/Sunbird-Ed/sb-cordova-plugin-db.git#release-3.1.0",
     "sb-cordova-plugin-sync": "git+https://github.com/project-sunbird/sb-cordova-plugin-sync.git#release-3.2.0",
@@ -113,7 +112,7 @@
     "sb-svg2pdf": "0.0.9",
     "tslib": "^1.10.0",
     "x2js": "^3.3.0",
-    "zone.js": "~0.8.29"
+    "zone.js": "~0.9.1"
   },
   "config": {
     "ionic_source_map_type": "inline-source-map",
@@ -122,14 +121,11 @@
     "ionic_enable_lint": false
   },
   "devDependencies": {
-    "@angular-devkit/architect": "0.13.9",
-    "@angular-devkit/build-angular": "0.13.9",
-    "@angular-devkit/core": "~7.3.8",
-    "@angular-devkit/schematics": "~7.3.8",
-    "@angular/cli": "~7.3.8",
-    "@angular/compiler": "~7.2.2",
-    "@angular/compiler-cli": "~7.2.2",
-    "@angular/language-service": "~7.2.2",
+    "@angular-devkit/build-angular": "^0.801.3",
+    "@angular/cli": "~8.3.29",
+    "@angular/compiler": "~8.0.0",
+    "@angular/compiler-cli": "~8.0.0",
+    "@angular/language-service": "~8.0.0",
     "@angularclass/hmr": "^2.1.3",
     "@babel/core": "^7.5.5",
     "@babel/plugin-proposal-class-properties": "^7.5.5",
@@ -138,7 +134,7 @@
     "@babel/plugin-transform-runtime": "^7.5.5",
     "@babel/preset-env": "^7.9.5",
     "@babel/preset-typescript": "^7.9.0",
-    "@ionic/angular-toolkit": "~1.5.1",
+    "@ionic/angular-toolkit": "^2.0.0",
     "@ionic/v4-migration-tslint": "^1.7.1",
     "@project-sunbird/telemetry-sdk": "0.0.26",
     "@types/jest": "^25.2.1",
@@ -150,7 +146,7 @@
     "sonarqube-scanner": "^2.5.0",
     "ts-node": "^8.4.1",
     "tslint": "~5.16.0",
-    "typescript": "~3.1.6"
+    "typescript": "~3.4.3"
   },
   "description": "An Ionic project",
   "cordova": {
