@@ -53,7 +53,7 @@ export class PlayerPage implements OnInit, OnDestroy, PlayerActionHandlerDelegat
   public objRollup: Rollup;
 
 
-  @ViewChild('preview') previewElement: ElementRef;
+  @ViewChild('preview', { static: false }) previewElement: ElementRef;
   constructor(
     @Inject('COURSE_SERVICE') private courseService: CourseService,
     private canvasPlayerService: CanvasPlayerService,
