@@ -59,7 +59,7 @@ import { AppVersion } from '@ionic-native/app-version/ngx';
 import { SbProgressLoader } from '@app/services/sb-progress-loader.service';
 import { FileOpener } from '@ionic-native/file-opener/ngx';
 import { TranslateService } from '@ngx-translate/core';
-import { FieldConfig } from 'common-form-elements';
+import { FieldConfig } from 'common-form-elements-v8';
 import { CertificateDownloadAsPdfService } from 'sb-svg2pdf';
 import { NavigationService } from '@app/services/navigation-handler.service';
 import { ContentUtil } from '@app/util/content-util';
@@ -75,7 +75,7 @@ import { ProfileHandler } from '@app/services/profile-handler';
 })
 export class ProfilePage implements OnInit {
 
-  @ViewChild('refresher') refresher: IonRefresher;
+  @ViewChild('refresher', { static: false }) refresher: IonRefresher;
 
   profile: any = {};
   userId = '';
