@@ -24,7 +24,7 @@ import {
   FilterValue,
   SearchType
 } from 'sunbird-sdk';
-import { LibraryCardTypes } from '@project-sunbird/common-consumption';
+import { LibraryCardTypes } from '@project-sunbird/common-consumption-v8';
 import { AppGlobalService, AppHeaderService, CommonUtilService, TelemetryGeneratorService } from '@app/services';
 import { animate, group, state, style, transition, trigger } from '@angular/animations';
 import { TranslateService } from '@ngx-translate/core';
@@ -83,7 +83,7 @@ import { CsPrimaryCategory } from '@project-sunbird/client-services/services/con
 export class ExploreBooksPage implements OnInit, OnDestroy {
   public pageId = 'ExploreBooksPage';
 
-  @ViewChild('searchInput') public searchInputRef: ElementRef;
+  @ViewChild('searchInput', { static: false }) public searchInputRef: ElementRef;
   @ViewChildren('filteredItems') public filteredItemsQueryList: QueryList<any>;
 
   LibraryCardTypes = LibraryCardTypes;
