@@ -8,7 +8,7 @@ import { IonicModule } from '@ionic/angular';
 import { DistrictMappingPage } from './district-mapping.page';
 import { TranslateModule } from '@ngx-translate/core';
 import { PipesModule } from '@app/pipes/pipes.module';
-import { CommonFormElementsModule } from 'common-form-elements';
+import { CommonFormElementsModule } from 'common-form-elements-v8';
 import { LocationHandler } from '@app/services/location-handler';
 import {ProfileHandler} from '@app/services/profile-handler';
 
