@@ -8,9 +8,9 @@ import { IonicModule } from '@ionic/angular';
 import { FaqReportIssuePage } from './faq-report-issue.page';
 import { TranslateModule } from '@ngx-translate/core';
 import { DirectivesModule } from '@app/directives/directives.module';
-import { CommonFormElementsModule } from 'common-form-elements';
+import { CommonFormElementsModule } from 'common-form-elements-v8';
 import { ComponentsModule } from '../components/components.module';
-import { CommonConsumptionModule } from '@project-sunbird/common-consumption';
+import { CommonConsumptionModule } from '@project-sunbird/common-consumption-v8';
 
 const routes: Routes = [
   {
