@@ -8,7 +8,7 @@ import { IonicModule } from '@ionic/angular';
 import { PlayerPage } from './player.page';
 import { CanvasPlayerService } from '@app/services/canvas-player.service';
 import { ScreenOrientation } from '@ionic-native/screen-orientation/ngx';
-import { SunbirdPdfPlayerModule } from '@project-sunbird/sunbird-pdf-player';
+import { SunbirdPdfPlayerModule } from '@project-sunbird/sunbird-pdf-player-v8';
 
 
 
