@@ -7,7 +7,7 @@ import { IonicModule } from '@ionic/angular';
 import { ComponentsModule } from '../components/components.module';
 import { DirectivesModule } from '@app/directives/directives.module';
 import { ReactiveFormsModule } from '@angular/forms';
-import { CommonConsumptionModule } from '@project-sunbird/common-consumption';
+import { CommonConsumptionModule } from '@project-sunbird/common-consumption-v8';
 import { ExploreBooksSortComponent } from './explore-books-sort/explore-books-sort.component';
 import { ResourcesRoutingModule } from './resources-routing.module';
 
